feat(integrations): Prep for login-with-VSTS
- Sets the identity's external id to the user ID
- Moves AccountConfigView() from the identity pipeline into the integrations pipeline
- Adds a build_identity method that will be used by the login flow
